Excel Exportation Report

+++ src/Api/PersistenceModels/PagoModel.cs
@@ -6,9 +6,9 @@
 
         public int CantCuota { get; set; }
 
-        public decimal Monto { get; set; }
+        public int Monto { get; set; }
 
-        public long NroRecibo { get; set; }
+        public int NroRecibo { get; set; }
 
         public DateTime FechaCarga { get; set; }
